# src/apiClients/codegen/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/todo/":{"get":{"summary":"Get Todos","operationId":"get_todos","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todos"}}}}}},"post":{"summary":"Create Todo","operationId":"create_todo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/todo/todo_id":{"put":{"summary":"Update Todo","operationId":"update_todo","parameters":[{"required":true,"schema":{"title":"Todo Id","type":"integer"},"name":"todo_id","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Todo","operationId":"delete_todo","parameters":[{"required":true,"schema":{"title":"Todo Id","type":"integer"},"name":"todo_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health_check/":{"get":{"summary":"Health Check","operationId":"health_check","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheck"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"HealthCheck":{"title":"HealthCheck","required":["status"],"type":"object","properties":{"status":{"title":"Status","type":"string"}}},"Todo":{"title":"Todo","required":["title","complete"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"title":{"title":"Title","type":"string"},"complete":{"title":"Complete","type":"boolean"}}},"Todos":{"title":"Todos","required":["todos"],"type":"object","properties":{"todos":{"title":"Todos","type":"array","items":{"$ref":"#/components/schemas/Todo"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}